디렉터일 경우 입력 비활성화, 의뢰서 관련 기능 구현 완료
