+++ lyc_flutter_project/lib/commissions/screen/basic_info_screen.dart
@@ -1,461 +1,431 @@
 import 'package:flutter/material.dart';
-import 'package:lyc_flutter_project/commissions/model/commission_model.dart';
-import 'package:lyc_flutter_project/common/widget/normal_appbar.dart';
+import 'package:lyc_flutter_project/common/widget/custom_loading.dart';
+import 'package:lyc_flutter_project/common/widget/custom_number_picker.dart';
 import 'package:lyc_flutter_project/data/app_color.dart';
+import 'package:lyc_flutter_project/data/style_list.dart' as styles;
+import 'package:lyc_flutter_project/setting/widget/custom_text_form_field.dart';
+import 'package:lyc_flutter_project/common/widget/default_padding.dart';
+import 'package:lyc_flutter_project/mypage/widget/select_buttons_in_posting.dart';
 import 'package:provider/provider.dart';
-import '../../../widget/Controller.dart';
-import '../../data/style_list.dart' as styles;
-import '../../styles/app_text_style.dart';
+import '../../auth/join/Provider/login_provider.dart';
 import '../model/commission_response_model.dart';
 import '../provider/commissions_provider.dart';
-import 'desired_style_screen.dart';
 
 class BasicInfoScreen extends StatefulWidget {
-  const BasicInfoScreen({super.key, required this.directorId, this.model});
+  const BasicInfoScreen(
+      {super.key, required this.model, required this.isDirector});
 
-  final int? directorId;
   final CommissionResponseModel? model;
-
+  final bool isDirector;
 
   @override
   State<BasicInfoScreen> createState() => BasicInfoScreenState();
 }
 
 class BasicInfoScreenState extends State<BasicInfoScreen> {
-  @override
-  void initState() {
-    super.initState();
-    final model = widget.model;
-
-    if (model != null) {
-      _controller1_1.controller.text = model.commissionInfo.height as String;
-      _controller1_2.controller.text = model.commissionInfo.weight as String;
-      _controller1_3.controller.text = model.commissionInfo.topSize;
-      _controller1_4.controller.text = model.commissionInfo.bottomSize;
-    }
-  }
-
-  final Controller _controller1_1 = Controller();
-  final Controller _controller1_2 = Controller();
-  final Controller _controller1_3 = Controller();
-  final Controller _controller1_4 = Controller();
-  final Controller _controller2 = Controller();
-  final Controller _controller3 = Controller();
-  final Controller _controller4 = Controller();
-  final Controller _controller5 = Controller();
-  final Controller _controller6 = Controller();
-  final Controller _controller7 = Controller();
-  final Controller _controller8 = Controller();
-  final Controller _controller9 = Controller();
-  final Controller _controller10 = Controller();
-
-
-  final membershipState = MembershipState();
-
-
+  int? memberId;
 
   @override
   Widget build(BuildContext context) {
-    Future<void> createCommissions() async {
-      final commissionsProvider =
-      Provider.of<CommissionsProvider>(context, listen: false);
-      commissionsProvider.model  = commissionsProvider.model.copyWith(
-          directorId: widget.directorId,
-          basicInfo: BasicInfo(
-              height: int.tryParse(_controller1_1.controller.text) ?? 0,
-              weight: int.tryParse(_controller1_2.controller.text) ?? 0,
-              topSize: _controller1_3.controller.text,
-              bottomSize: _controller1_4.controller.text,
-
-              text: _controller10.controller.text,
-              infoStyle: InfoStyle(
-                  preferredStyleList: _controller2.controller.text.split(','),
-                  nonPreferredStyleList:
-                      _controller3.controller.text.split(',')),
-              infoFit: InfoFit(
-                  preferredFitList: _controller6.controller.text.split(','),
-                  nonPreferredFitList: _controller7.controller.text.split(',')),
-              infoMaterial: InfoMaterial(
-                  preferredMaterialList:
-                      _controller4.controller.text.split(','),
-                  nonPreferredMaterialList:
-                      _controller5.controller.text.split(',')),
-              infoBodyType: InfoBodyType(
-                  goodBodyTypeList: _controller8.controller.text.split(','),
-                  badBodyTypeList: _controller9.controller.text.split(','))));
-    }
-
-    return ChangeNotifierProvider(
-      create: (_) => membershipState,
-      child: Builder(
-        builder: (context) => Scaffold(
-          backgroundColor: AppColor.lightGrey,
-          appBar: NormalAppbar(
-            title: widget.title,
-          ),
-          body: SingleChildScrollView(
-            child: Center(
-              child: Container(
-                margin: const EdgeInsets.fromLTRB(20, 16, 20, 16),
-                child: Column(
-                  crossAxisAlignment: CrossAxisAlignment.center,
-                  children: [
-                    Padding(
-                      padding: const EdgeInsets.only(bottom: 20),
-                      child: buildBodyTypeSection(),
-                    ),
-                    Padding(
-                      padding: const EdgeInsets.only(bottom: 20),
-                      child: buildStyleSection(
-                        title: 'Step 2. 평소 즐겨입는 스타일은 무엇인가요?',
-                        notifier: membershipState._selected2,
-                        styles: styles.StyleList.styleOptions,
-                        text: '기타 스타일 입력',
-                        controller: _controller2,
-                      ),
-                    ),
-                    Padding(
-                      padding: const EdgeInsets.only(bottom: 20),
-                      child: buildStyleSection(
-                        title: 'Step 3. 평소 즐겨입지 않는 스타일은 무엇인가요?',
-                        notifier: membershipState._selected3,
-                        styles: styles.StyleList.styleOptions,
-                        text: '기타 스타일 입력',
-                        controller: _controller3,
-                      ),
-                    ),
-                    Padding(
-                      padding: const EdgeInsets.only(bottom: 20),
-                      child: buildStyleSection(
-                        title: 'Step 4. 선호하는 소재를 선택해 주세요.',
-                        notifier: membershipState._selected4,
-                        styles: styles.StyleList.materialOptions,
-                        text: '기타 소재 입력',
-                        controller: _controller4,
-                      ),
-                    ),
-                    Padding(
-                      padding: const EdgeInsets.only(bottom: 20),
-                      child: buildStyleSection(
-                        title: 'Step 5. 선호하지 않는 소재를 선택해 주세요.',
-                        notifier: membershipState._selected5,
-                        styles: styles.StyleList.materialOptions,
-                        text: '기타 소재 입력',
-                        controller: _controller5,
-                      ),
-                    ),
-                    Padding(
-                      padding: const EdgeInsets.only(bottom: 20),
-                      child: buildStyleSection(
-                        title: 'Step 6. 선호하는 핏을 선택해 주세요.',
-                        notifier: membershipState._selected6,
-                        styles: styles.StyleList.fitOptions,
-                        text: '기타 핏 입력',
-                        controller: _controller6,
-                      ),
-                    ),
-                    Padding(
-                      padding: const EdgeInsets.only(bottom: 20),
-                      child: buildStyleSection(
-                        title: 'Step 7. 선호하지 않는 핏을 선택해 주세요.',
-                        notifier: membershipState._selected7,
-                        styles: styles.StyleList.fitOptions,
-                        text: '기타 핏 입력',
-                        controller: _controller7,
-                      ),
-                    ),
-                    Padding(
-                      padding: const EdgeInsets.only(bottom: 20),
-                      child: buildStyleSection(
-                        title: 'Step 8. 강조하고 싶은 신체 부위가 있나요?',
-                        notifier: membershipState._selected8,
-                        styles: styles.StyleList.bodyParts,
-                        text: '기타 핏 입력',
-                        controller: _controller8,
-                      ),
-                    ),
-                    Padding(
-                      padding: const EdgeInsets.only(bottom: 20),
-                      child: buildStyleSection(
-                        title: 'Step 9. 보완하고 싶은 신체 부위가 있나요?',
-                        notifier: membershipState._selected9,
-                        styles: styles.StyleList.bodyParts,
-                        text: '기타 핏 입력',
-                        controller: _controller9,
-                      ),
-                    ),
-                    Padding(
-                      padding: const EdgeInsets.only(bottom: 20),
-                      child:
-                          buildAdditionalInfo('Step 10.추가로 작성하고 싶은 내용이 있나요?'),
-
-                    ),
-                    Padding(
-                        padding: const EdgeInsets.only(bottom: 20),
-                        child: TextButton(
-                          onPressed: () async {
-                            try {
-                              await createCommissions();
-                            } catch (e) {
-                              debugPrint("저장 실패-1");
-                            }
-                          },
-                          style: TextButton.styleFrom(
-                            backgroundColor: AppColor.brown, // 색상 변경
-                            minimumSize: const Size(120, 40), // 버튼 크기 설정
-                            shape: RoundedRectangleBorder(
-                              borderRadius: BorderRadius.circular(20),
+    final bool isDirector = widget.isDirector;
+    return Scaffold(
+      backgroundColor: AppColor.lightGrey,
+      body: Consumer<LoginProvider>(
+        builder: (context, loginProvider, child) {
+          memberId = loginProvider.memberId;
+          return DefaultPadding(
+            bottom: 20.0,
+            child: Consumer<CommissionsProvider>(
+              builder: (context, value, child) {
+                int topSize = int.parse(value.model.basicInfo.topSize);
+                int bottomSize = int.parse(value.model.basicInfo.bottomSize);
+                if (memberId == null) {
+                  return const Center(
+                    child: CustomLoading(),
+                  );
+                } else {
+                  return ListView(
+                    keyboardDismissBehavior:
+                        ScrollViewKeyboardDismissBehavior.onDrag,
+                    children: [
+                      ContentBox(
+                        title: "1. 본인의 체형을 알려주세요.",
+                        child: Column(
+                          children: [
+                            SpecInputLine(
+                                label: "키",
+                                initialValue:
+                                    value.model.basicInfo.height.toString(),
+                                onChanged: (p0) =>
+                                    value.updateHeight(height: int.parse(p0)),
+                                enabled:
+                                !isDirector),
+                            SpecInputLine(
+                                label: "몸무게",
+                                initialValue:
+                                    value.model.basicInfo.weight.toString(),
+                                onChanged: (p0) =>
+                                    value.updateWeight(weight: int.parse(p0)),
+                                enabled:    !isDirector),
+                            SpecInputLine(
+                              label: "상의 사이즈",
+                              initialValue: "",
+                              value: value.model.basicInfo.topSize,
+                              enabled:    !isDirector,
+                              onTap: () {
+                                showDialog(
+                                  context: context,
+                                  builder: (context) =>
+                                      CustomNumberPicker<CommissionsProvider>(
+                                    title: "상의 사이즈를 선택해주세요.",
+                                    minValue: 80,
+                                    maxValue: 120,
+                                    getValue: (p0) =>
+                                        int.parse(p0.model.basicInfo.topSize),
+                                    step: 5,
+                                    updateValue: (p0, p1) =>
+                                        p0.updateTopSize(topSize: p1),
+                                    fstOnPressed: () {
+                                      value.updateTopSize(topSize: topSize);
+                                      Navigator.pop(context);
+                                    },
+                                    scdOnPressed: () => Navigator.pop(context),
+                                  ),
+                                );
+                              },
                             ),
-                          ),
-                          child: const Text(
-                            '저장',
-                            style: AppTextStyle.button,
-                            textAlign: TextAlign.center,
-                          ),
-                        ))
-                  ],
-                ),
-              ),
+                            SpecInputLine(
+                              label: "하의 사이즈",
+                              initialValue: "",
+                              value: value.model.basicInfo.bottomSize,
+                              enabled: !isDirector,
+                              onTap: () {
+                                showDialog(
+                                  context: context,
+                                  builder: (context) =>
+                                      CustomNumberPicker<CommissionsProvider>(
+                                    title: "하의 사이즈를 선택해주세요.",
+                                    minValue: 24,
+                                    maxValue: 42,
+                                    getValue: (p0) => int.parse(
+                                        p0.model.basicInfo.bottomSize),
+                                    updateValue: (p0, p1) =>
+                                        p0.updateBottomSize(bottomSize: p1),
+                                    fstOnPressed: () {
+                                      value.updateBottomSize(
+                                          bottomSize: bottomSize);
+                                      Navigator.pop(context);
+                                    },
+                                    scdOnPressed: () => Navigator.pop(context),
+                                  ),
+                                );
+                              },
+                            ),
+                          ],
+                        ),
+                      ),
+                      ContentBox(
+                        title: "2. 평소 즐겨입는 스타일은 무엇인가요?",
+                        child: ButtonList(
+                            name: styles.StyleList.styleOptions,
+                            selected: value
+                                .model.basicInfo.infoStyle.preferredStyleList,
+                            onSelected: (v) =>
+                                value.updatePreferredStyle(selected: v),
+                            enabled:    !isDirector),
+                      ),
+                      ContentBox(
+                        title: "3. 평소 즐겨입지 않는 스타일은 무엇인가요?",
+                        child: ButtonList(
+                            name: styles.StyleList.styleOptions,
+                            selected: value.model.basicInfo.infoStyle
+                                .nonPreferredStyleList,
+                            onSelected: (v) =>
+                                value.updateNonPreferredStyle(selected: v),
+                            enabled:    !isDirector),
+                      ),
+                      ContentBox(
+                        title: "4. 선호하는 소재를 선택해주세요.",
+                        child: ButtonList(
+                            name: styles.StyleList.materialOptions,
+                            selected: value.model.basicInfo.infoMaterial
+                                .preferredMaterialList,
+                            onSelected: (v) =>
+                                value.updatePreferredMaterials(selected: v),
+                            enabled:    !isDirector),
+                      ),
+                      ContentBox(
+                        title: "5. 선호하지 않는 소재를 선택해주세요.",
+                        child: ButtonList(
+                            name: styles.StyleList.materialOptions,
+                            selected: value.model.basicInfo.infoMaterial
+                                .nonPreferredMaterialList,
+                            onSelected: (v) =>
+                                value.updateNonPreferredMaterials(selected: v),
+                            enabled:    !isDirector),
+                      ),
+                      ContentBox(
+                        title: "6. 선호하는 핏을 선택해주세요.",
+                        child: ButtonList(
+                            name: styles.StyleList.fitOptions,
+                            selected:
+                                value.model.basicInfo.infoFit.preferredFitList,
+                            onSelected: (v) =>
+                                value.updatePreferredFits(selected: v),
+                            enabled:    !isDirector),
+                      ),
+                      ContentBox(
+                        title: "7. 선호하지 않는 핏을 선택해주세요.",
+                        child: ButtonList(
+                            name: styles.StyleList.fitOptions,
+                            selected: value
+                                .model.basicInfo.infoFit.nonPreferredFitList,
+                            onSelected: (v) =>
+                                value.updateNonPreferredFits(selected: v),
+                            enabled:    !isDirector),
+                      ),
+                      ContentBox(
+                        title: "9. 강조하고 싶은 신체 부위가 있나요?",
+                        child: ButtonList(
+                            name: styles.StyleList.bodyParts,
+                            selected: value
+                                .model.basicInfo.infoBodyType.goodBodyTypeList,
+                            onSelected: (v) =>
+                                value.updateBadBodyTypes(selected: v),
+                            enabled:    !isDirector),
+                      ),
+                      ContentBox(
+                        title: "10. 보완하고 싶은 신체 부위가 있나요?",
+                        child: ButtonList(
+                            name: styles.StyleList.bodyParts,
+                            selected: value
+                                .model.basicInfo.infoBodyType.badBodyTypeList,
+                            onSelected: (v) =>
+                                value.updateBadBodyTypes(selected: v),
+                            enabled:    !isDirector),
+                      ),
+                      ContentBox(
+                        title: "10. 추가로 작성하고 싶은 내용이 있나요?",
+                        child: CustomTextFormField(
+                            hint: "ex. 종아리가 너무 두꺼운 게 고민이에요.",
+                            maxLines: 5,
+                            containerMargin: 0.0,
+                            focusedBorderColor: Colors.transparent,
+                            focusedBorderWidth: 0.0,
+                            contentPaddingHorizontal: 8.0,
+                            initialValue: value.model.basicInfo.text,
+                            onChanged: (text) => value.updateText(text: text),
+                            enabled: !isDirector),
+                      ),
+                    ],
+                  );
+                }
+              },
             ),
-          ),
-        ),
+          );
+        },
       ),
     );
   }
+}
 
-  Widget buildAdditionalInfo(String text) {
-    return Container(
-      width: 332,
-      height: 200,
-      decoration: BoxDecoration(
-        color: Colors.white,
-        borderRadius: BorderRadius.circular(20),
-      ),
-      padding: const EdgeInsets.fromLTRB(20, 16, 20, 16),
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Text(
-            text,
-            style: AppTextStyle.littleTitle.copyWith(fontSize: 14),
-          ),
-          const SizedBox(height: 10),
-          Expanded(
-              child: TextField(
-            controller: _controller10.controller,
-            maxLines: null,
-            decoration: InputDecoration(
-              contentPadding:
-                  const EdgeInsets.symmetric(vertical: 0, horizontal: 12),
-              filled: false,
-              hintText: 'ex.종아리가 너무 두꺼운 게 고민이에요.',
-              border: OutlineInputBorder(
-                borderRadius: BorderRadius.circular(20),
-                borderSide: BorderSide.none,
-              ),
-            ),
-            style: AppTextStyle.hint.copyWith(color: Colors.black),
-          )),
-        ],
-      ),
-    );
-  }
+class ButtonList extends StatelessWidget {
+  final List<String> name;
+  final List<String> selected;
+  final Function(String) onSelected;
+  final bool enabled;
 
-  Widget buildBodyTypeSection() {
-    return Container(
-      width: 332,
-      decoration: BoxDecoration(
-        color: Colors.white,
-        borderRadius: BorderRadius.circular(20),
-      ),
-      padding: const EdgeInsets.fromLTRB(20, 16, 20, 16),
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Center(
-            child: Padding(
-              padding: const EdgeInsets.only(bottom: 12),
-              child: Text(
-                'Step 1. 본인의 체형을 알려주세요.',
-                style: AppTextStyle.littleTitle.copyWith(fontSize: 14.0),
-              ),
-            ),
-          ),
-          Align(
-            alignment: Alignment.centerRight,
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.end,
-              children: [
-                buildInputField('키', _controller1_1),
-                buildInputField('몸무게', _controller1_2),
-                buildInputField('상의 사이즈', _controller1_3),
-                buildInputField('하의 사이즈', _controller1_4),
-              ],
-            ),
-          ),
-        ],
-      ),
-    );
-  }
+  const ButtonList(
+      {super.key,
+      required this.name,
+      required this.selected,
+      required this.onSelected,
+      required this.enabled});
 
-  Widget buildInputField(String label, Controller controller) {
-    return Container(
-      margin: const EdgeInsets.only(bottom: 8),
-      child: Row(
-        children: [
-          Expanded(
-            flex: 1,
-            child: Align(
-              alignment: Alignment.centerRight,
-              child: Text(
-                label,
-                style: AppTextStyle.hint.copyWith(color: Colors.black),
-              ),
-            ),
-          ),
-          const SizedBox(width: 12),
-          Expanded(
-            flex: 2,
-            child: SizedBox(
-              height: 20,
-              child: TextField(
-                controller: controller.controller,
-                decoration: InputDecoration(
-                  contentPadding:
-                      const EdgeInsets.symmetric(vertical: 0, horizontal: 0),
-                  fillColor: AppColor.grey,
-                  filled: true,
-                  border: OutlineInputBorder(
-                    borderRadius: BorderRadius.circular(20),
-                    borderSide: BorderSide.none,
-                  ),
-                ),
-                style: AppTextStyle.hint.copyWith(color: Colors.black),
-                keyboardType: TextInputType.number,
-              ),
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-
-  Widget buildStyleSection({
-    required String title,
-    required ValueNotifier<List<String>> notifier,
-    required List<String> styles,
-    required String text,
-    required Controller controller,
-  }) {
-    notifier.addListener(() {
-      controller.controller.text = notifier.value.join(',');
-    });
-    return Container(
-      width: 332,
-      height: 212,
-      decoration: BoxDecoration(
-        color: Colors.white,
-        borderRadius: BorderRadius.circular(20),
-      ),
-      padding: const EdgeInsets.fromLTRB(16, 20, 16, 12),
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Padding(
-              padding: const EdgeInsets.only(bottom: 12),
-              child: Text(
-                title,
-                style: AppTextStyle.littleTitle.copyWith(fontSize: 14.0),
-              )),
-          Padding(
-              padding: const EdgeInsets.only(bottom: 12),
-              child: buildStyleOptions(styles: styles, notifier: notifier)),
-          SizedBox(
-            width: 296,
-            height: 30,
-            child: TextField(
-              controller: controller.controller,
-              decoration: InputDecoration(
-                hintText: text,
-                filled: true,
-                fillColor: AppColor.grey,
-                border: OutlineInputBorder(
-                  borderRadius: BorderRadius.circular(20),
-                  borderSide: BorderSide.none,
-                ),
-                contentPadding:
-                    const EdgeInsets.symmetric(horizontal: 12, vertical: 12),
-              ),
-              style: AppTextStyle.hint.copyWith(color: Colors.black),
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-
-  Widget buildStyleOptions({
-    required List<String> styles,
-    required ValueNotifier<List<String>> notifier,
-  }) {
-    final List<String> styles1 = styles.sublist(0, 4);
-    final List<String> styles2 = styles.sublist(4);
-
+  @override
+  Widget build(BuildContext context) {
     return Column(
       children: [
         Row(
-          mainAxisAlignment: MainAxisAlignment.spaceBetween,
-          children: styles1
-              .map((style) => buildStyleButton(style, notifier))
-              .toList(),
+          children: [
+            for (var i = 0; i < 4; i++)
+              Expanded(
+                child: SelectButtonsInPosting(
+                  name,
+                  selected,
+                  i,
+                  () => onSelected(name[i]),
+                  AppColor.brown,
+                  AppColor.lightGrey,
+                 enabled: enabled,
+                ),
+              ),
+          ],
         ),
-        const SizedBox(height: 6),
         Row(
-          mainAxisAlignment: MainAxisAlignment.spaceBetween,
-          children: styles2
-              .map((style) => buildStyleButton(style, notifier))
-              .toList(),
-        ),
+          children: [
+            for (var i = 4; i < 8; i++)
+              Expanded(
+                child: SelectButtonsInPosting(
+                  name,
+                  selected,
+                  i,
+                  () => onSelected(name[i]),
+                  AppColor.brown,
+                  AppColor.lightGrey,
+                 enabled: !enabled,
+                ),
+              ),
+          ],
+        )
       ],
     );
   }
+}
 
-  Widget buildStyleButton(String style, ValueNotifier<List<String>> notifier) {
-    return ValueListenableBuilder<List<String>>(
-      valueListenable: notifier,
-      builder: (context, selectedStyles, child) {
-        final isSelected = selectedStyles.contains(style);
-        return Expanded(
-          child: GestureDetector(
-            onTap: () {
-              final updatedStyles = List<String>.from(selectedStyles);
-              if (isSelected) {
-                updatedStyles.remove(style);
-              } else {
-                updatedStyles.add(style);
-              }
-              notifier.value = updatedStyles;
-            },
-            child: Container(
-              margin: const EdgeInsets.only(right: 4),
-              decoration: BoxDecoration(
-                color: isSelected ? AppColor.brown : AppColor.lightGrey,
-                borderRadius: BorderRadius.circular(20),
+class ContentBox extends StatelessWidget {
+  final String title;
+  final Widget child;
+
+  const ContentBox({
+    super.key,
+    required this.title,
+    required this.child,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      margin: const EdgeInsets.only(
+        bottom: 20.0,
+      ),
+      padding: const EdgeInsets.symmetric(
+        vertical: 16.0,
+        horizontal: 20.0,
+      ),
+      //height: 200.0,
+      decoration: BoxDecoration(
+        color: Colors.white,
+        borderRadius: BorderRadius.circular(
+          20.0,
+        ),
+      ),
+      child: Column(
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: [
+          Padding(
+            padding: const EdgeInsets.all(8.0),
+            child: Text(
+              title,
+              style: const TextStyle(
+                fontWeight: FontWeight.w600,
+                fontSize: 16.0,
               ),
-              padding: const EdgeInsets.symmetric(vertical: 8),
-              child: Center(
-                child: Text(
-                  style,
-                  style: isSelected
-                      ? const TextStyle(
-                          fontWeight: FontWeight.w500,
-                          fontSize: 14,
-                          color: AppColor.lightGrey)
-                      : const TextStyle(
-                          fontWeight: FontWeight.w500,
-                          fontSize: 14,
-                          color: Colors.black),
-                ),
+            ),
+          ),
+          child,
+        ],
+      ),
+    );
+  }
+}
+
+class SpecInputLine extends StatelessWidget {
+  final String label;
+  final String? initialValue;
+  final VoidCallback? onTap;
+  final String? value;
+  final Function(String)? onChanged;
+  final bool enabled;
+
+  const SpecInputLine({
+    super.key,
+    required this.label,
+    required this.initialValue,
+    this.onTap,
+    this.value,
+    this.onChanged,
+    required this.enabled,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return Padding(
+      padding: const EdgeInsets.symmetric(horizontal: 20.0),
+      child: Row(
+        crossAxisAlignment: CrossAxisAlignment.center,
+        children: [
+          Expanded(
+            child: Text(
+              label,
+              style: const TextStyle(
+                fontWeight: FontWeight.w500,
+                fontSize: 16.0,
+              ),
+              textAlign: TextAlign.end,
+            ),
+          ),
+          const SizedBox(
+            width: 20.0,
+          ),
+          Expanded(
+            child: (label == "키" || label == "몸무게")
+                ? CustomTextFormField(
+                    fillColor: const Color(0xffE9E9E9),
+                    focusedBorderColor: Colors.black,
+                    focusedBorderWidth: 1.5,
+                    contentPaddingVertical: 4.0,
+                    fontSize: 16.0,
+                    isDense: true,
+                    initialValue: initialValue ?? '',
+                    keyboardType: TextInputType.number,
+                    onChanged: onChanged!,
+                    enabled: enabled,
+                  )
+                : SpecSizeBox(
+                    onTap: onTap!,
+                    string: value,
+                  ),
+          ),
+        ],
+      ),
+    );
+  }
+}
+
+class SpecSizeBox extends StatelessWidget {
+  final VoidCallback onTap;
+  final String? string;
+
+  const SpecSizeBox({
+    super.key,
+    required this.onTap,
+    required this.string,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return Consumer<CommissionsProvider>(
+      builder: (context, value, child) {
+        return GestureDetector(
+          onTap: onTap,
+          child: Container(
+            margin: const EdgeInsets.symmetric(vertical: 8.0),
+            decoration: BoxDecoration(
+              color: const Color(0xffE9E9E9),
+              borderRadius: BorderRadius.circular(20.0),
+            ),
+            width: double.infinity,
+            alignment: Alignment.centerLeft,
+            padding: const EdgeInsets.symmetric(
+              vertical: 4.0,
+              horizontal: 20.0,
+            ),
+            child: Text(
+              string!,
+              style: const TextStyle(
+                fontSize: 16.0,
+                fontWeight: FontWeight.w500,
               ),
             ),
           ),
@@ -463,27 +433,4 @@ class BasicInfoScreenState extends State<BasicInfoScreen> {
       },
     );
   }
-
-
-
-
-}
-
-class MembershipState extends ChangeNotifier {
-  final ValueNotifier<List<String>> _selected2 =
-      ValueNotifier<List<String>>([]);
-  final ValueNotifier<List<String>> _selected3 =
-      ValueNotifier<List<String>>([]);
-  final ValueNotifier<List<String>> _selected4 =
-      ValueNotifier<List<String>>([]);
-  final ValueNotifier<List<String>> _selected5 =
-      ValueNotifier<List<String>>([]);
-  final ValueNotifier<List<String>> _selected6 =
-      ValueNotifier<List<String>>([]);
-  final ValueNotifier<List<String>> _selected7 =
-      ValueNotifier<List<String>>([]);
-  final ValueNotifier<List<String>> _selected8 =
-      ValueNotifier<List<String>>([]);
-  final ValueNotifier<List<String>> _selected9 =
-      ValueNotifier<List<String>>([]);
 }
