+++ lyc_flutter_project/lib/setting/screen/style_screen.dart
@@ -33,7 +33,6 @@ class _StyleScreenState extends State<StyleScreen> {
       body: Consumer<LoginProvider>(
         builder: (context, loginProvider, child) {
           memberId = loginProvider.memberId;
-
           return DefaultPadding(
             bottom: 20.0,
             child: Consumer<SettingProvider>(
